# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * XQution - Physics and Astronomy Formula Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Formula {
  id: number;
  name: string;
  category: string;
  /** @nullable */
  subcategory?: string | null;
  latex: string;
  description: string;
  variables?: string;
  /** @nullable */
  siUnits?: string | null;
  /** @nullable */
  example?: string | null;
  /** @nullable */
  relatedFormulas?: string | null;
  isFeatured?: boolean;
}

export interface Constant {
  id: number;
  name: string;
  symbol: string;
  value: string;
  units: string;
  description: string;
  /** @nullable */
  category?: string | null;
}

export interface Problem {
  id: number;
  topic: string;
  question: string;
  difficulty: string;
  /** @nullable */
  hint?: string | null;
  /** @nullable */
  solution?: string | null;
  /** @nullable */
  answer?: string | null;
}

export interface GlossaryTerm {
  id: number;
  term: string;
  definition: string;
  /** @nullable */
  category?: string | null;
}

export interface Favorite {
  id: number;
  itemType: string;
  itemId: number;
  itemName: string;
  createdAt?: string;
}

export interface FavoriteInput {
  itemType: string;
  itemId: number;
  itemName: string;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface SearchResults {
  formulas: Formula[];
  constants: Constant[];
  problems: Problem[];
  glossary: GlossaryTerm[];
}

export interface PlatformStats {
  formulaCount: number;
  constantCount: number;
  problemCount: number;
  glossaryCount: number;
  categoryCount: number;
}

export type ListFormulasParams = {
category?: string;
search?: string;
};

export type ListConstantsParams = {
search?: string;
};

export type ListProblemsParams = {
topic?: string;
difficulty?: string;
search?: string;
};

export type ListGlossaryTermsParams = {
search?: string;
};

export type GlobalSearchParams = {
q: string;
};