# lib/api-zod/src/generated/types/favorite.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * XQution - Physics and Astronomy Formula Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Favorite {
  id: number;
  itemType: string;
  itemId: number;
  itemName: string;
  createdAt?: string;
}