# lib/api-zod/src/generated/types/problem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * XQution - Physics and Astronomy Formula Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Problem {
  id: number;
  topic: string;
  question: string;
  difficulty: string;
  /** @nullable */
  hint?: string | null;
  /** @nullable */
  solution?: string | null;
  /** @nullable */
  answer?: string | null;
}